+++ src/test/java/cat/udl/eps/entsoftarch/webingogeiapi/steps/DeleteGameStepDefs.java
@@ -19,7 +19,7 @@ public class DeleteGameStepDefs {
     private StepDefs stepDefs;
 
     @When("^I delete a game with id \"([^\"]*)\"$")
-    public void iDeleteAGameWithId(int id) throws Throwable {
+    public void iDeleteAGameWithId(long id) throws Throwable {
         stepDefs.result = stepDefs.mockMvc.perform(
                 delete("/games/{id}", id)
                         .contentType(MediaType.APPLICATION_JSON)
@@ -29,7 +29,7 @@ public class DeleteGameStepDefs {
     }
 
     @And("^It has not been deleted a game with id \"([^\"]*)\"$")
-    public void itHasNotBeenDeletedAGameWithId(int id) throws Throwable {
+    public void itHasNotBeenDeletedAGameWithId(long id) throws Throwable {
         stepDefs.result = stepDefs.mockMvc.perform(
                 get("/games/{id}", id)
                         .accept(MediaType.APPLICATION_JSON))
@@ -38,7 +38,7 @@ public class DeleteGameStepDefs {
 
 
     @And("^It has been deleted a game with id \"([^\"]*)\"$")
-    public void itHasBeenDeletedAGameWithId(int id) throws Throwable {
+    public void itHasBeenDeletedAGameWithId(long id) throws Throwable {
         stepDefs.result = stepDefs.mockMvc.perform(
                 get("/games/{id}", id)
                         .accept(MediaType.APPLICATION_JSON)
